requirements.txt
aiogram>=3.4,<4
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9,<3
Jinja2>=3.1,<4
python-multipart>=0.0.9,<1
itsdangerous>=2.1,<3
python-dotenv>=1,<2
httpx>=0.27,<1
